# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login com e-mail ou CPF e senha.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/select-context": {
      "post": {
        "operationId": "AuthController_selectContext",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectContextDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Seleciona contexto de acesso e emite JWT de sessão",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Cadastro via código de convite",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/join-context": {
      "post": {
        "operationId": "AuthController_joinContext",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinContextDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Usuário existente aceita convite e vincula novo contexto",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/request-password": {
      "post": {
        "operationId": "AuthController_requestPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Solicita e-mail para redefinir senha (e-mail ou CPF)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Redefine senha com token recebido por e-mail",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Perfil do usuário autenticado (JWT)",
        "tags": [
          "auth"
        ]
      }
    },
    "/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar empresas",
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar empresa",
        "tags": [
          "companies"
        ]
      }
    },
    "/companies/{id}/users": {
      "get": {
        "operationId": "CompaniesController_companyUsers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar vínculos da empresa (usuários)",
        "tags": [
          "companies"
        ]
      }
    },
    "/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe da empresa",
        "tags": [
          "companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar empresa",
        "tags": [
          "companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inativar empresa (soft delete)",
        "tags": [
          "companies"
        ]
      }
    },
    "/companies/{id}/invite-links": {
      "post": {
        "operationId": "CompaniesController_invite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCompanyInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gerar link de convite",
        "tags": [
          "companies"
        ]
      }
    },
...
    "/company-users": {
      "get": {
        "operationId": "CompanyUsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar usuários da empresa com mapa de permissões por operador",
        "tags": [
          "company-users"
        ]
      }
    },
    "/company-users/invite-links": {
      "get": {
        "operationId": "CompanyUsersController_listInviteLinks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar convites ativos da empresa",
        "tags": [
          "company-users"
        ]
      },
      "post": {
        "operationId": "CompanyUsersController_generateInviteLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCompanyUserInviteLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gerar convite para novo membro da empresa",
        "tags": [
          "company-users"
        ]
      }
    },
    "/company-users/{companyUserId}/role": {
      "patch": {
        "operationId": "CompanyUsersController_updateRole",
        "parameters": [
          {
            "name": "companyUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompanyUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alterar papel do colaborador",
        "tags": [
          "company-users"
        ]
      }
    },
    "/company-users/{companyUserId}/active": {
      "patch": {
        "operationId": "CompanyUsersController_setActive",
        "parameters": [
          {
            "name": "companyUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompanyUserActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ativar/desativar colaborador",
        "tags": [
          "company-users"
        ]
      }
    },
    "/company-users/{companyUserId}/profile": {
      "patch": {
        "operationId": "CompanyUsersController_updateProfile",
        "parameters": [
          {
            "name": "companyUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompanyUserProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar nome exibido e dados do vínculo",
        "tags": [
          "company-users"
        ]
      }
    },
    "/company-users/{companyUserId}/permissions": {
      "patch": {
        "operationId": "CompanyUsersController_updatePermissions",
        "parameters": [
          {
            "name": "companyUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCompanyUserPermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Permissões granulares por módulo (operadores)",
        "tags": [
          "company-users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Face2go API",
    "description": "Backend NestJS da plataforma Face2go",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "identifier": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "password"
        ]
      },
      "SelectContextDto": {
        "type": "object",
        "properties": {
          "contextType": {
            "type": "string",
            "enum": [
              "super_admin",
              "company",
              "client",
              "responsible",
              "member",
              "face_user"
            ]
          },
          "contextId": {
            "type": "string"
          }
        },
        "required": [
          "contextType"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255
          },
          "phone": {
            "type": "string",
            "minLength": 8,
            "maxLength": 30
          },
          "jobTitle": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "invite": {
            "type": "string",
            "minLength": 4
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "invite"
        ]
      },
      "JoinContextDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "invite": {
            "type": "string",
            "minLength": 4
          }
        },
        "required": [
          "identifier",
          "password",
          "invite"
        ]
      },
      "RequestPasswordDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "identifier"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255
          },
          "cnpj": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "isActive"
        ]
      },
      "PatchCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255
          },
          "cnpj": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "GenerateCompanyInviteDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "role": {
            "type": "string",
            "enum": [
              "company_admin",
              "company_operator"
            ]
          }
        },
        "required": [
          "companyId",
          "role"
        ]
      },
      "GenerateCompanyUserInviteLinkDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "company_admin",
              "company_operator"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "PatchCompanyUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "company_admin",
              "company_operator"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "PatchCompanyUserActiveDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "isActive"
        ]
      },
      "PatchCompanyUserProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255
          },
          "jobTitle": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "nullable": true
          },
          "phone": {
            "type": "string",
            "minLength": 8,
            "maxLength": 30,
            "nullable": true
          }
        }
      },
      "PatchCompanyUserPermissionsDto": {
        "type": "object",
        "properties": {
          "featureSlug": {
            "type": "string"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "can_read",
                "can_create",
                "can_update",
                "can_delete"
              ]
            }
          }
        },
        "required": [
          "featureSlug",
          "actions"
        ]
      },
...
    }
  }
}